refactor: remove the `slot.rs`
The naming is not proper. (Slot) `NotAssigned` and `Assigned` are much
more proper.
