+++ kernel/src/device/pci/xhci/port/mod.rs
@@ -2,15 +2,18 @@
 
 use super::{
     exchanger,
-    structures::{context::Context, registers},
+    structures::{context::Context, dcbaa, descriptor, registers},
 };
 use crate::multitask::{self, task::Task};
-use alloc::collections::VecDeque;
+use alloc::{collections::VecDeque, sync::Arc, vec::Vec};
 use conquer_once::spin::Lazy;
 use core::{future::Future, pin::Pin, task::Poll};
+use descriptor::Descriptor;
+use endpoint::Endpoint;
+use exchanger::{transfer, transfer::DoorbellWriter};
 use futures_util::task::AtomicWaker;
+use page_box::PageBox;
 use resetter::Resetter;
-use slot::Slot;
 use spinning_top::Spinlock;
 use xhci::registers::operational::PortRegisterSet;
 
@@ -18,7 +21,6 @@ mod class_driver;
 mod context;
 mod endpoint;
 mod resetter;
-mod slot;
 mod spawner;
 
 static CURRENT_RESET_PORT: Lazy<Spinlock<ResetPort>> =
@@ -58,7 +60,7 @@ pub fn try_spawn(port_idx: u8) -> Result<(), spawner::PortNotConnected> {
     spawner::try_spawn(port_idx)
 }
 
-async fn task(port: Port) {
+async fn task(port: SlotNotAssigned) {
     let mut eps = init_port_and_slot_exclusively(port).await;
     eps.init().await;
 
@@ -74,7 +76,7 @@ async fn task(port: Port) {
     }
 }
 
-async fn init_port_and_slot_exclusively(port: Port) -> endpoint::Collection {
+async fn init_port_and_slot_exclusively(port: SlotNotAssigned) -> endpoint::Collection {
     let reset_waiter = ResetWaiterFuture;
     reset_waiter.await;
 
@@ -85,13 +87,13 @@ async fn init_port_and_slot_exclusively(port: Port) -> endpoint::Collection {
     endpoint::Collection::new(slot).await
 }
 
-async fn init_port_and_slot(mut p: Port) -> Slot {
+async fn init_port_and_slot(mut p: SlotNotAssigned) -> SlotAssigned {
     p.reset();
     p.init_context();
 
     let slot_id = exchanger::command::enable_device_slot().await;
 
-    let mut slot = Slot::new(p, slot_id);
+    let mut slot = SlotAssigned::new(p, slot_id);
     slot.init().await;
     debug!("Slot initialized");
     slot
@@ -105,11 +107,11 @@ fn max_num() -> u8 {
     registers::handle(|r| r.capability.hcsparams1.read().number_of_ports())
 }
 
-pub struct Port {
+pub struct SlotNotAssigned {
     index: u8,
     context: Context,
 }
-impl Port {
+impl SlotNotAssigned {
     fn new(index: u8) -> Self {
         Self {
             index,
@@ -133,6 +135,139 @@ impl Port {
 
     fn read_port_rg(&self) -> PortRegisterSet {
         registers::handle(|r| r.port_register_set.read_at((self.index - 1).into()))
+    }
+}
+
+pub struct SlotAssigned {
+    id: u8,
+    cx: Arc<Spinlock<Context>>,
+    def_ep: endpoint::Default,
+}
+impl SlotAssigned {
+    pub fn new(port: SlotNotAssigned, id: u8) -> Self {
+        let cx = Arc::new(Spinlock::new(port.context));
+        let dbl_writer = DoorbellWriter::new(id, 1);
+        Self {
+            id,
+            cx: cx.clone(),
+            def_ep: endpoint::Default::new(transfer::Sender::new(dbl_writer), cx, port.index),
+        }
+    }
+
+    pub(in crate::device::pci::xhci) fn context(&self) -> Arc<Spinlock<Context>> {
+        self.cx.clone()
+    }
+
+    pub fn id(&self) -> u8 {
+        self.id
+    }
+
+    pub async fn init(&mut self) {
+        self.init_default_ep();
+        self.register_with_dcbaa();
+        self.issue_address_device().await;
+    }
+
+    pub async fn get_device_descriptor(&mut self) -> PageBox<descriptor::Device> {
+        self.def_ep.get_device_descriptor().await
+    }
+
+    pub async fn endpoints(&mut self) -> Vec<Endpoint> {
+        let ds = self.get_configuration_descriptors().await;
+        let mut eps = Vec::new();
+
+        for d in ds {
+            if let Descriptor::Endpoint(ep) = d {
+                eps.push(self.generate_endpoint(ep));
+            }
+        }
+
+        eps
+    }
+
+    pub async fn interface_descriptor(&mut self) -> descriptor::Interface {
+        *self
+            .get_configuration_descriptors()
+            .await
+            .iter()
+            .find_map(|x| {
+                if let Descriptor::Interface(e) = x {
+                    Some(e)
+                } else {
+                    None
+                }
+            })
+            .unwrap()
+    }
+
+    pub async fn get_configuration_descriptors(&mut self) -> Vec<Descriptor> {
+        let r = self.get_raw_configuration_descriptors().await;
+        RawDescriptorParser::new(r).parse()
+    }
+
+    fn generate_endpoint(&self, ep: descriptor::Endpoint) -> Endpoint {
+        Endpoint::new(
+            ep,
+            self.cx.clone(),
+            transfer::Sender::new(DoorbellWriter::new(self.id, ep.doorbell_value())),
+        )
+    }
+
+    fn init_default_ep(&mut self) {
+        self.def_ep.init_context();
+    }
+
+    async fn get_raw_configuration_descriptors(&mut self) -> PageBox<[u8]> {
+        self.def_ep.get_raw_configuration_descriptors().await
+    }
+
+    fn register_with_dcbaa(&mut self) {
+        let a = self.cx.lock().output.phys_addr();
+        dcbaa::register(self.id.into(), a);
+    }
+
+    async fn issue_address_device(&mut self) {
+        let cx_addr = self.cx.lock().input.phys_addr();
+        exchanger::command::address_device(cx_addr, self.id).await;
+    }
+}
+
+struct RawDescriptorParser {
+    raw: PageBox<[u8]>,
+    current: usize,
+    len: usize,
+}
+impl RawDescriptorParser {
+    fn new(raw: PageBox<[u8]>) -> Self {
+        let len = raw.len();
+        Self {
+            raw,
+            current: 0,
+            len,
+        }
+    }
+
+    fn parse(&mut self) -> Vec<Descriptor> {
+        let mut v = Vec::new();
+        while self.current < self.len && self.raw[self.current] > 0 {
+            match self.parse_first_descriptor() {
+                Ok(t) => v.push(t),
+                Err(e) => debug!("Unrecognized USB descriptor: {:?}", e),
+            }
+        }
+        v
+    }
+
+    fn parse_first_descriptor(&mut self) -> Result<Descriptor, descriptor::Error> {
+        let raw = self.cut_raw_descriptor();
+        Descriptor::from_slice(&raw)
+    }
+
+    fn cut_raw_descriptor(&mut self) -> Vec<u8> {
+        let len: usize = self.raw[self.current].into();
+        let v = self.raw[self.current..(self.current + len)].to_vec();
+        self.current += len;
+        v
     }
 }
