+++ kernel/src/device/pci/xhci/port/spawner.rs
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-use super::Port;
+use super::SlotNotAssigned;
 use crate::multitask;
 use alloc::{vec, vec::Vec};
 use conquer_once::spin::Lazy;
@@ -18,7 +18,7 @@ pub(in crate::device::pci::xhci) fn spawn_all_connected_ports() {
 }
 
 pub(in crate::device::pci::xhci) fn try_spawn(port_idx: u8) -> Result<(), PortNotConnected> {
-    let p = Port::new(port_idx);
+    let p = SlotNotAssigned::new(port_idx);
     if spawnable(&p) {
         spawn(p);
         Ok(())
@@ -27,24 +27,24 @@ pub(in crate::device::pci::xhci) fn try_spawn(port_idx: u8) -> Result<(), PortNo
     }
 }
 
-fn spawn(p: Port) {
+fn spawn(p: SlotNotAssigned) {
     mark_as_spawned(&p);
     add_task_for_port(p);
 }
 
-fn add_task_for_port(p: Port) {
+fn add_task_for_port(p: SlotNotAssigned) {
     multitask::add(Task::new(super::task(p)));
 }
 
-fn spawnable(p: &Port) -> bool {
+fn spawnable(p: &SlotNotAssigned) -> bool {
     p.connected() && !spawned(p)
 }
 
-fn spawned(p: &Port) -> bool {
+fn spawned(p: &SlotNotAssigned) -> bool {
     SPAWN_STATUS.lock()[usize::from(p.index)]
 }
 
-fn mark_as_spawned(p: &Port) {
+fn mark_as_spawned(p: &SlotNotAssigned) {
     SPAWN_STATUS.lock()[usize::from(p.index)] = true;
 }
